refactor(93): reorganise HiveMQ theme as a proper module
