+++ hivemq-edge/src/frontend/src/modules/App/MainApp.tsx
@@ -2,8 +2,9 @@ import { FC } from 'react'
 import { RouterProvider } from 'react-router-dom'
 import { ChakraProvider } from '@chakra-ui/react'
 
+import { themeHiveMQ } from '@/modules/Theme/themeHiveMQ.ts'
+
 import { routes } from './routes.tsx'
-import { themeHiveMQ } from './themes/themeHiveMQ.ts'
 import { AuthProvider } from '../Auth/AuthProvider.tsx'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
