+++ hivemq-edge/src/frontend/src/modules/Theme/styles/Stat.ts
